user-service and account-service updated and working

+++ user-service/src/main/java/fon/bank/userservice/dao/ClientRepository.java
@@ -4,8 +4,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import fon.bank.userservice.entity.Client;
 
+import java.util.Optional;
+
 @Repository
-public interface ClientRepository extends JpaRepository<Client, Integer> {
-//    Client findClientByUserClientUsername(String username);
+public interface ClientRepository extends JpaRepository<Client, Long> {
       Client findByUserClient(String username);
 }
